# apps/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BestMatchRequest": {
        "additionalProperties": false,
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/PostingFilters"
          },
          "profile_text": {
            "default": "",
            "maxLength": 50000,
            "title": "Profile Text",
            "type": "string"
          },
          "query": {
            "default": "",
            "maxLength": 500,
            "title": "Query",
            "type": "string"
          }
        },
        "title": "BestMatchRequest",
        "type": "object"
      },
      "CatalogueQueryRequest": {
        "additionalProperties": false,
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/PostingFilters"
          },
          "page": {
            "default": 1,
            "maximum": 9007199254740991.0,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "query": {
            "default": "",
            "maxLength": 500,
            "title": "Query",
            "type": "string"
          },
          "sort": {
            "$ref": "#/components/schemas/CatalogueSort",
            "default": "newest"
          }
        },
        "title": "CatalogueQueryRequest",
        "type": "object"
      },
      "CatalogueSort": {
        "enum": [
          "newest",
          "salary"
        ],
        "title": "CatalogueSort",
        "type": "string"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "error",
          "meta"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "MetaData": {
        "additionalProperties": false,
        "properties": {
          "corpus_size": {
            "minimum": 0.0,
            "title": "Corpus Size",
            "type": "integer"
          },
          "retrieval": {
            "title": "Retrieval",
            "type": "string"
          },
          "rewrite_provider": {
            "title": "Rewrite Provider",
            "type": "string"
          },
          "source_counts": {
            "items": {
              "$ref": "#/components/schemas/SourceCountData"
            },
            "title": "Source Counts",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceId"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "corpus_size",
          "sources",
          "source_counts",
          "retrieval",
          "rewrite_provider"
        ],
        "title": "MetaData",
        "type": "object"
      },
...
      "PostingLookupRequest": {
        "additionalProperties": false,
        "properties": {
          "ids": {
            "items": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "title": "PostingLookupRequest",
        "type": "object"
      },
...
      "RankingStage": {
        "enum": [
          "rewrite",
          "filter",
          "retrieve",
          "group",
          "rerank"
        ],
        "title": "RankingStage",
        "type": "string"
      },
...
      "SearchCompleted": {
        "additionalProperties": false,
        "properties": {
          "event": {
            "const": "search.completed",
            "default": "search.completed",
            "title": "Event",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/BestMatchData"
          },
          "took_ms": {
            "minimum": 0.0,
            "title": "Took Ms",
            "type": "number"
          }
        },
        "required": [
          "request_id",
          "snapshot",
          "took_ms"
        ],
        "title": "SearchCompleted",
        "type": "object"
      },
      "SearchFailed": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "event": {
            "const": "search.failed",
            "default": "search.failed",
            "title": "Event",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "error"
        ],
        "title": "SearchFailed",
        "type": "object"
      },
      "SearchStarted": {
        "additionalProperties": false,
        "properties": {
          "event": {
            "const": "search.started",
            "default": "search.started",
            "title": "Event",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "SearchStarted",
        "type": "object"
      },
...
      "StageCompleted": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "duration_ms": {
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "number"
          },
          "event": {
            "const": "stage.completed",
            "default": "stage.completed",
            "title": "Event",
            "type": "string"
          },
          "item_count": {
            "minimum": 0.0,
            "title": "Item Count",
            "type": "integer"
          },
          "ordinal": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Ordinal",
            "type": "integer"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "stage": {
            "$ref": "#/components/schemas/RankingStage"
          },
          "status": {
            "$ref": "#/components/schemas/TraceStatus"
          }
        },
        "required": [
          "request_id",
          "stage",
          "ordinal",
          "status",
          "detail",
          "item_count",
          "duration_ms"
        ],
        "title": "StageCompleted",
        "type": "object"
      },
      "StageStarted": {
        "additionalProperties": false,
        "properties": {
          "event": {
            "const": "stage.started",
            "default": "stage.started",
            "title": "Event",
            "type": "string"
          },
          "ordinal": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Ordinal",
            "type": "integer"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "stage": {
            "$ref": "#/components/schemas/RankingStage"
          }
        },
        "required": [
          "request_id",
          "stage",
          "ordinal"
        ],
        "title": "StageStarted",
        "type": "object"
      },
      "SuccessResponse_BestMatchData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BestMatchData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[BestMatchData]",
        "type": "object"
      },
      "SuccessResponse_MetaData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MetaData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[MetaData]",
        "type": "object"
      },
      "SuccessResponse_PostingDetail_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PostingDetail"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[PostingDetail]",
        "type": "object"
      },
      "SuccessResponse_list_PostingSummary__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PostingSummary"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[list[PostingSummary]]",
        "type": "object"
      },
      "SuccessResponse_list_ResolvedPosting__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ResolvedPosting"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[list[ResolvedPosting]]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Search API over the RAG pipeline",
    "title": "jobber",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/meta": {
      "get": {
        "operationId": "meta_api_meta_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_MetaData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Meta"
      }
    },
    "/api/postings/lookup": {
      "post": {
        "operationId": "posting_lookup_api_postings_lookup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostingLookupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_list_ResolvedPosting__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Posting Lookup"
      }
    },
    "/api/postings/query": {
      "post": {
        "operationId": "query_postings_api_postings_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogueQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_list_PostingSummary__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Query Postings"
      }
    },
    "/api/postings/{posting_id}": {
      "get": {
        "operationId": "posting_api_postings__posting_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "posting_id",
            "required": true,
            "schema": {
              "maxLength": 512,
              "title": "Posting Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_PostingDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Posting"
      }
    },
    "/api/search": {
      "post": {
        "operationId": "search_api_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BestMatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_BestMatchData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Search"
      }
    },
    "/api/search/stream": {
      "post": {
        "operationId": "search_stream_api_search_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BestMatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "properties": {
                    "data": {
                      "contentMediaType": "application/json",
                      "contentSchema": {
                        "discriminator": {
                          "mapping": {
                            "search.completed": "#/components/schemas/SearchCompleted",
                            "search.failed": "#/components/schemas/SearchFailed",
                            "search.started": "#/components/schemas/SearchStarted",
                            "stage.completed": "#/components/schemas/StageCompleted",
                            "stage.started": "#/components/schemas/StageStarted"
                          },
                          "propertyName": "event"
                        },
                        "oneOf": [
                          {
                            "$ref": "#/components/schemas/SearchStarted"
                          },
                          {
                            "$ref": "#/components/schemas/StageStarted"
                          },
                          {
                            "$ref": "#/components/schemas/StageCompleted"
                          },
                          {
                            "$ref": "#/components/schemas/SearchCompleted"
                          },
                          {
                            "$ref": "#/components/schemas/SearchFailed"
                          }
                        ],
                        "title": "Streamitem Search Stream Api Search Stream Post"
                      },
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "minimum": 0,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Search Stream"
      }
    }
  }
}